requirements.txt
google-genai
yfinance
python-dotenv
